Update position before momentum so collision behavior works
momentums and positions are both calculated inside Character:walk().
But collisions are calculated by DoCharacterCollisions() in the outer loop.
Collisions work zeroing momentums in blocked directions.
So if momentum is calculated before position in walk(), momentum can
carry characters in blocked directions before the collisions are checked.
Had to fix that!